+++ R/app_Run.R
@@ -7,31 +7,33 @@
 #' @export
 #' @importFrom shiny shinyApp
 #' @importFrom golem with_golem_options
-run_app <- function(pathToCohortOperationsConfigYalm, pathToDatabasesConfigYalm, ...) {
+run_app <- function(pathToDatabasesConfigYalm, pathToAnalysisModulesConfigYalm, ...) {
 
-  # set up configuration
-  checkmate::assertFileExists(pathToCohortOperationsConfigYalm, extension = "yml")
-  cohortOperationsConfig <- yaml::read_yaml(pathToCohortOperationsConfigYalm)
-  checkmate::assertList(cohortOperationsConfig, names = "named")
-
+  # Check configuration files
+  # TODO: check if the config files are correct
   checkmate::assertFileExists(pathToDatabasesConfigYalm, extension = "yml")
   databasesConfig <- yaml::read_yaml(pathToDatabasesConfigYalm)
-  checkmate::assertList(databasesConfig, names = "named")
 
+  checkmate::assertFileExists(pathToAnalysisModulesConfigYalm, extension = "yml")
+  analysisModulesConfig <- yaml::read_yaml(pathToAnalysisModulesConfigYalm)
+
+  #
+  # GLOBAL SETTING
+  #
 
   # set shiny to accept large files
-  options(shiny.maxRequestSize = 1000000000)
+  options(shiny.maxRequestSize = 10*1000*1024^2) # 10GB
 
   # deactivate https request to work with Atlas in https
   httr::set_config(httr::config(ssl_verifypeer = FALSE))
 
-  # set up futuress
-  future::plan(future::multisession, workers = 2)
-
   # set up loger
-  logger <- fcr_setUpLogger()
+  fcr_setUpLogger()
 
-  # start app
+  #
+  # Create app
+  #
+
   app  <- shiny::shinyApp(
     ui = app_ui,
     server = app_server,
@@ -39,31 +41,19 @@ run_app <- function(pathToCohortOperationsConfigYalm, pathToDatabasesConfigYalm,
   )
 
   # setup shiny options
-  app$appOptions$cohortOperationsConfig  <- cohortOperationsConfig
   app$appOptions$databasesConfig  <- databasesConfig
-  # TEMP
-  app$appOptions$logger  <- logger
-  app$appOptions$cores  <- parallel::detectCores()-1
-  app$appOptions$chunksSizeNOutcomes  <- 1000
+  app$appOptions$analysisModulesConfig  <- analysisModulesConfig
 
   app$appOptions$pathToNews  <- here::here("NEWS.md")
   app$appOptions$gitInfo  <- paste("Branch: ", gert::git_info()$shorthand, "Commit: ", gert::git_info()$commit)
 
+  #
+  # Launch app
+  #
+
+  # log start
   ParallelLogger::logInfo("[Start] Start logging on ", app$appOptions$gitInfo)
 
   return(app)
 }
 
-
-.createConsoleAppenderForSandboxLogging <- function(layout = ParallelLogger::layoutParallel) {
-  appendFunction <- function(this, level, message, echoToConsole) {
-    # Avoid note in check:
-    missing(this)
-    message <- paste0("[sandbox-co2-log] ", message)
-    writeLines(message, con = stdout())
-
-  }
-  appender <- list(appendFunction = appendFunction, layout = layout)
-  class(appender) <- "Appender"
-  return(appender)
-}