+++ R/app_Server.R
@@ -5,87 +5,50 @@
 #' @import shiny
 #' @noRd
 app_server <- function(input, output, session) {
-  # Your application server logic
-
   # get settings loaded from file
   databasesConfig <- shiny::getShinyOption("databasesConfig")
 
-  # get connection sandbox API configured for running GWAS
-  connection_sandboxAPI <- NULL
-  tryCatch({
-    connection_sandboxAPI <- configGWAS()
-  }, error=function(e) {
-    ParallelLogger::logError("[configGWAS]: ", e$message)
-  }, warning=function(w) {
-    ParallelLogger::logWarn("[configGWAS]: ", w$message)
-  })
-
-  # list of connection handlers that are passed to modules,
-  # not all modules used all, they modules check the list has at least the ones they need
-  # they are produced only by mod_select_configuration and consumed by the modules
-  r_connectionHandlers <- shiny::reactiveValues(
-    databasesHandlers = NULL,
-    connection_sandboxAPI = connection_sandboxAPI
+  # This is a reactive value with cohortTableHandler and a counter to indicate a change with in the object
+  # Because the cohortTableHandler is a pointe to an object, it changes only when the the selected database changes, but changes with in the object are not reflected
+  # To comunicate changes with in the object, we use a counter: hasChangeCounter
+  # Modules that modify information with in the object should increment this counter
+  r_connectionHandler <- shiny::reactiveValues(
+    cohortTableHandler = NULL,
+    hasChangeCounter = 0
   )
 
-  # produced by modules related to cohort editing
-  # consumed by cohort table viewer
-  r_workbench <- shiny::reactiveValues(
-    cohortsSummaryDatabases = HadesExtras::createEmptyCohortsSummary()
-  )
 
-  shiny::observe({
-    sweetAlert_spinner("Starting application")
-  })
+  mod_selectDatabases_server("selectDatabases", databasesConfig, r_connectionHandler)
 
-  mod_selectDatabases_server("selectDatabases", databasesConfig, r_connectionHandlers)
+  mod_cohortWorkbench_server("cohortWorkbench_importCohorts", r_connectionHandler)
+  mod_importCohortsFromFile_server("importCohortsFromFile", r_connectionHandler)
+  mod_importCohortsFromAtlas_server("importCohortsFromAtlas", r_connectionHandler)
+  mod_importCohortsFromCohortsTable_server("importCohortsFromEndpoints", r_connectionHandler)
+  mod_importCohortsFromAtlas_server("importCohortsFromLibrary", r_connectionHandler, filterCohortsRegex = ".*\\[CohortLibrary]")
 
-  mod_cohortWorkbench_server("cohortWorkbench_importCohorts", r_connectionHandlers, r_workbench)
-  mod_importCohortsFromFile_server("importCohortsFromFile", r_connectionHandlers, r_workbench)
-  mod_importCohortsFromAtlas_server("importCohortsFromAtlas", r_connectionHandlers, r_workbench)
-  mod_importCohortsFromCohortsTable_server("importCohortsFromEndpoints", r_connectionHandlers, r_workbench)
-  mod_importCohortsFromAtlas_server("importCohortsFromLibrary", r_connectionHandlers, r_workbench, filterCohortsRegex = ".*\\[CohortLibrary]")
+  mod_cohortWorkbench_server("cohortWorkbench_matchCohorts", r_connectionHandler)
+  mod_matchCohorts_server("matchCohorts", r_connectionHandler)
 
-  mod_cohortWorkbench_server("cohortWorkbench_matchCohorts", r_connectionHandlers, r_workbench)
-  mod_matchCohorts_server("matchCohorts", r_connectionHandlers, r_workbench)
+  mod_cohortWorkbench_server("cohortWorkbench_operateCohorts", r_connectionHandler)
+  mod_operateCohorts_server("operateCohorts", r_connectionHandler)
 
-  mod_cohortWorkbench_server("cohortWorkbench_operateCohorts", r_connectionHandlers, r_workbench)
-  mod_operateCohorts_server("operateCohorts", r_connectionHandlers, r_workbench)
+  mod_cohortWorkbench_server("cohortWorkbench_exportsCohorts", r_connectionHandler)
+  mod_exportsCohorts_server("exportsCohorts", r_connectionHandler)
 
-  mod_cohortWorkbench_server("cohortWorkbench_exportsCohorts", r_connectionHandlers, r_workbench)
-  mod_exportsCohorts_server("exportsCohorts", r_connectionHandlers, r_workbench)
+  # Dynamic analysis modules server
+  lapply(names(analysisModulesConfig), function(analysisKey) {
+    analysis <- analysisModulesConfig[[analysisKey]]
+    mod_cohortWorkbench_server(paste0("cohortWorkbench_", analysisKey), r_connectionHandler)
 
-  mod_cohortWorkbench_server("cohortWorkbench_cohortDiagnostics", r_connectionHandlers, r_workbench)
-  mod_cohortDiagnostics_server("cohortDiagnostics", r_connectionHandlers, r_workbench)
-
-  mod_cohortWorkbench_server("cohortWorkbench_cohortOverlaps", r_connectionHandlers, r_workbench)
-  mod_cohortOverlaps_server("cohortOverlaps", r_connectionHandlers, r_workbench)
-
-  mod_cohortWorkbench_server("cohortWorkbench_cohortsIncidence", r_connectionHandlers, r_workbench)
-  mod_cohortsIncidence_server("cohortsIncidence", r_connectionHandlers, r_workbench)
-
-  mod_cohortWorkbench_server("cohortWorkbench_timeCodeWAS", r_connectionHandlers, r_workbench)
-  mod_timeCodeWAS_server("timeCodeWAS", r_connectionHandlers, r_workbench)
-
-  mod_cohortWorkbench_server("cohortWorkbench_codeWAS", r_connectionHandlers, r_workbench)
-  mod_codeWAS_server("codeWAS", r_connectionHandlers, r_workbench)
-
-  mod_cohortWorkbench_server("cohortWorkbench_runGWAS", r_connectionHandlers, r_workbench)
-  mod_runGWAS_server("runGWAS", r_connectionHandlers, r_workbench)
-
-
-  output$about <- shiny::renderUI({
-    # load news from shinyoption pathtomd
-
-    shiny::div(
-      shiny::markdown(readLines(shiny::getShinyOption("pathToNews"))),
-      shiny::br(),
-      shiny::p(shiny::getShinyOption("gitInfo"))
+    mod_analysisWrap_server(
+      id = paste0(analysisKey, "_analysis"),
+      r_connectionHandler = r_connectionHandler,
+      mod_analysisSettings_server = analysis$mod_analysisSettings_server |> fct_stringToFuction(),
+      fct_executeAnalysis = analysis$fct_executeAnalysis |> fct_stringToFuction(),
+      analysisName = analysis$analysisName,
+      url_visualiseResults = analysis$url_visualiseResults
     )
-
   })
-
-
 
 }