+++ R/mod_OperateCohorts.R
@@ -13,29 +13,18 @@
 mod_operateCohorts_ui <- function(id) {
   ns <- shiny::NS(id)
   htmltools::tagList(
-    mod_appendCohort_ui(),
+    mod_fct_appendCohort_ui(),
     shinyjs::useShinyjs(),
-    #
-    shiny::tags$h4("Database"),
-    shiny::tags$h5("Select database where to work:"),
-    shiny::tags$h6("(Only cohort in the same database can be matched)"),
-    shinyWidgets::pickerInput(
-      inputId = ns("selectDatabases_pickerInput"),
-      label = NULL,
-      choices = NULL,
-      selected = NULL,
-      multiple = FALSE),
     htmltools::hr(),
     shiny::tags$h4("Operation - expression defining the cohort"),
-    #shiny::textInput(inputId = ns("operationString_textInput"), label = "Operation String"),
     htmltools::div(style = "width: 90%; height: 260; overflow: auto; margin-left: 30px; padding: 10px;",
-      mod_dragAndDrop_ui(ns("dragAndDrop"))
+                   mod_dragAndDrop_ui(ns("dragAndDrop"))
     ),
     #
     htmltools::hr(),
     shiny::tags$h4("Cohort intersections - UpSet plot"),
     htmltools::div(style = "width: 90%; height: 300; overflow: auto; margin-left: 30px; padding: 10px;",
-      shiny::plotOutput(ns("upsetPlot"), width = "80%", height = "250px"),
+                   shiny::plotOutput(ns("upsetPlot"), width = "80%", height = "250px"),
     ),
     htmltools::hr(),
     shiny::tags$h4("New cohort name"),
@@ -46,7 +35,7 @@ mod_operateCohorts_ui <- function(id) {
   )
 }
 
-mod_operateCohorts_server <- function(id, r_connectionHandlers, r_workbench) {
+mod_operateCohorts_server <- function(id, r_connectionHandler) {
   shiny::moduleServer(id, function(input, output, session) {
     ns <- session$ns
 
@@ -55,45 +44,32 @@ mod_operateCohorts_server <- function(id, r_connectionHandlers, r_workbench) {
     #
     r <- shiny::reactiveValues(
       cohortDefinitionSet = NULL,
-      operationStringError = NULL
+      operationStringError = NULL,
+      operationString = NULL
     )
 
-    r_toAdd <- shiny::reactiveValues(
-      databaseName = NULL,
+    r_cohortDefinitionSetToAdd <- shiny::reactiveValues(
       cohortDefinitionSet = NULL
     )
-
-
-    #
-    # update selectDatabases_pickerInput with database names
-    #
-    shiny::observe({
-      shiny::req(r_connectionHandlers$databasesHandlers)
-
-      databaseIdNamesList <- fct_getDatabaseIdNamesListFromDatabasesHandlers(r_connectionHandlers$databasesHandlers)
-
-      shinyWidgets::updatePickerInput(
-        inputId = "selectDatabases_pickerInput",
-        choices = databaseIdNamesList,
-        selected = databaseIdNamesList[1]
-      )
-    })
-
 
     #
     # get operation string from dragAndDrop module
     #
-    rf_operationString <- mod_dragAndDrop_server("dragAndDrop", r_workbench)
+    rf_operationString <- mod_dragAndDrop_server("dragAndDrop", r_connectionHandler)
+    # copy to reactive value for easier testing
+    shiny::observe(
+      r$operationString <- rf_operationString()
+    )
 
     #
     # when operation string changes check if it is valid
     #
     shiny::observe({
-      shiny::req(rf_operationString())
+      shiny::req(r$operationString)
 
       error <- NULL
       tryCatch({
-        HadesExtras::operationStringToSQL(rf_operationString())
+        HadesExtras::operationStringToSQL(r$operationString)
       }, error = function(e){
         error <<- e$message
       })
@@ -107,69 +83,69 @@ mod_operateCohorts_server <- function(id, r_connectionHandlers, r_workbench) {
     # if operation string is valid create temporal cohortDefinitionSet and render name
     #
     shiny::observe({
-      shiny::req(input$selectDatabases_pickerInput)
-      shiny::req(rf_operationString())
-      shiny::req(is.null(r$operationStringError))
+      shiny::req(r$operationString)
+      # if not valid set cohortDefinitionSet to NULL
+      if(!is.null(r$operationStringError)){
+        r$cohortDefinitionSet <- NULL
+      }else{
+        existingSubsetDefinitionIds <- r_connectionHandler$cohortTableHandler$cohortDefinitionSet |>
+          dplyr::filter(!is.na(subsetDefinitionId)) |>
+          dplyr::pull(subsetDefinitionId)
 
-      cohortTableHandler <- r_connectionHandlers$databasesHandlers[[input$selectDatabases_pickerInput]]$cohortTableHandler
+        nextSubsetDefinitionId <- ifelse( length(existingSubsetDefinitionIds) == 0, 1, max(existingSubsetDefinitionIds)+1 )
 
-      existingSubsetDefinitionIds <- cohortTableHandler$cohortDefinitionSet |>
-        dplyr::filter(!is.na(subsetDefinitionId)) |>
-        dplyr::pull(subsetDefinitionId)
-
-      nextSubsetDefinitionId <- ifelse( length(existingSubsetDefinitionIds) == 0, 1, max(existingSubsetDefinitionIds)+1 )
-
-      subsetDef <- CohortGenerator::createCohortSubsetDefinition(
-        name = "",
-        definitionId = nextSubsetDefinitionId,
-        subsetOperators = list(
-          HadesExtras::createOperationSubset(
-            operationString = rf_operationString()
+        subsetDef <- CohortGenerator::createCohortSubsetDefinition(
+          name = "",
+          definitionId = nextSubsetDefinitionId,
+          subsetOperators = list(
+            HadesExtras::createOperationSubset(
+              operationString = r$operationString
+            )
           )
         )
-      )
 
-      # TEMP::get first integer From operation string
-      targetCohortIds <- as.integer(stringr::str_extract(rf_operationString(), "\\d+"))
+        # TEMP::get first integer From operation string
+        targetCohortIds <- as.integer(stringr::str_extract(r$operationString, "\\d+"))
 
-      # two adjacent cohorts crash the following function (malformed input)
-      cohortDefinitionSet <- CohortGenerator::addCohortSubsetDefinition(
-        cohortDefinitionSet = cohortTableHandler$cohortDefinitionSet,
-        cohortSubsetDefintion = subsetDef,
-        targetCohortIds = targetCohortIds,
-        overwriteExisting =  TRUE
-      )
+        # two adjacent cohorts crash the following function (malformed input)
+        cohortDefinitionSet <- CohortGenerator::addCohortSubsetDefinition(
+          cohortDefinitionSet = r_connectionHandler$cohortTableHandler$cohortDefinitionSet,
+          cohortSubsetDefintion = subsetDef,
+          targetCohortIds = targetCohortIds,
+          overwriteExisting =  TRUE
+        )
 
-      cohortDefinitionSetOnlyNew <- cohortDefinitionSet |>
-        dplyr::filter(subsetDefinitionId == nextSubsetDefinitionId)
+        cohortDefinitionSetOnlyNew <- cohortDefinitionSet |>
+          dplyr::filter(subsetDefinitionId == nextSubsetDefinitionId)
 
-      # update cohortId to non existing, to avoid overflow here
-      # https://github.com/OHDSI/CohortGenerator/blob/e3efad630b8b2c0376431a88fde89e6c4bbac38c/R/SubsetDefinitions.R#L193
-      previousCohortId <- cohortDefinitionSetOnlyNew$cohortId
-      unusedCohortId <- setdiff(1:1000, cohortTableHandler$cohortDefinitionSet |> dplyr::pull(cohortId)) |> head(1)
+        # update cohortId to non existing, to avoid overflow here
+        # https://github.com/OHDSI/CohortGenerator/blob/e3efad630b8b2c0376431a88fde89e6c4bbac38c/R/SubsetDefinitions.R#L193
+        previousCohortId <- cohortDefinitionSetOnlyNew$cohortId
+        unusedCohortId <- setdiff(1:1000, r_connectionHandler$cohortTableHandler$cohortDefinitionSet |> dplyr::pull(cohortId)) |> head(1)
 
-      cohortDefinitionSetOnlyNew <- cohortDefinitionSetOnlyNew |>
-        dplyr::mutate(
-          cohortId = unusedCohortId,
-          sql = stringr::str_replace(sql, paste0('cohort_definition_id = ', previousCohortId), paste0('cohort_definition_id = ', unusedCohortId)),
-          sql = stringr::str_replace(sql, paste0(previousCohortId, ' as cohort_definition_id'), paste0(unusedCohortId, ' as cohort_definition_id'))
+        cohortDefinitionSetOnlyNew <- cohortDefinitionSetOnlyNew |>
+          dplyr::mutate(
+            cohortId = unusedCohortId,
+            sql = stringr::str_replace(sql, paste0('cohort_definition_id = ', previousCohortId), paste0('cohort_definition_id = ', unusedCohortId)),
+            sql = stringr::str_replace(sql, paste0(previousCohortId, ' as cohort_definition_id'), paste0(unusedCohortId, ' as cohort_definition_id'))
           )
 
-      r$cohortDefinitionSet <- cohortDefinitionSetOnlyNew
-
+        r$cohortDefinitionSet <- cohortDefinitionSetOnlyNew
+      }
     })
 
     #
     # Render temporal name
     #
     output$newCohortName_text <- shiny::renderText({
-      stringToShow <- "----"
+      if(is.null(r$cohortDefinitionSet) && is.null(r$operationStringError)){
+        stringToShow  <- "----"
+      }
       if(!is.null(r$operationStringError)){
-        stringToShow <- paste("❌", r$operationStringError)
-      }else{
-        if(shiny::isTruthy(r$cohortDefinitionSet)){
-          stringToShow <- paste("✅", r$cohortDefinitionSet$cohortName)
-        }
+        stringToShow  <- paste("❌", r$operationStringError)
+      }
+      if(!is.null(r$cohortDefinitionSet)){
+        stringToShow  <- paste("✅", r$cohortDefinitionSet$cohortName)
       }
 
       ParallelLogger::logInfo("[Operate cohorts] Operation: ", stringToShow)
@@ -181,16 +157,14 @@ mod_operateCohorts_server <- function(id, r_connectionHandlers, r_workbench) {
     # create upset plot
     #
     output$upsetPlot <- shiny::renderPlot({
-
-      if(!is.null(r$operationStringError)){
+      if(is.null(r$cohortDefinitionSet)){
         return(NULL)
       } else {
         if(shiny::isTruthy(r$cohortDefinitionSet)){
 
-          s <- rf_operationString()
+          s <- r$operationString
           cohortsInOperation <- as.character(stringr::str_extract_all(s, "\\d+")[[1]])
-          cohortTableHandler <- r_connectionHandlers$databasesHandlers[[input$selectDatabases_pickerInput]]$cohortTableHandler
-          cohortOverlap <- cohortTableHandler$getCohortsOverlap()
+          cohortOverlap <- r_connectionHandler$cohortTableHandler$getCohortsOverlap()
 
           # TEMP FIX
           # extract all numbers from string s
@@ -227,7 +201,7 @@ mod_operateCohorts_server <- function(id, r_connectionHandlers, r_workbench) {
             dplyr::group_by_if(is.logical) |>
             dplyr::summarise(numberOfSubjects = sum(numberOfSubjects), .groups = "drop")
 
-          .plot_upset_cohortsOverlap(cohortOverlap, cohortTableHandler$getCohortIdAndNames() |> dplyr::select(cohortId, shortName) )
+          .plot_upset_cohortsOverlap(cohortOverlap, r_connectionHandler$cohortTableHandler$getCohortIdAndNames() |> dplyr::select(cohortId, shortName) )
         }
       }
     })
@@ -247,11 +221,11 @@ mod_operateCohorts_server <- function(id, r_connectionHandlers, r_workbench) {
       shiny::req(r$cohortDefinitionSet)
 
       ## copy selected to
-      r_toAdd$databaseName <- input$selectDatabases_pickerInput
-      r_toAdd$cohortDefinitionSet <-  r$cohortDefinitionSet
+      r_cohortDefinitionSetToAdd$databaseName <- input$selectDatabases_pickerInput
+      r_cohortDefinitionSetToAdd$cohortDefinitionSet <-  r$cohortDefinitionSet
 
-      ParallelLogger::logInfo("[Operate cohorts] Creating cohorts: ", r_toAdd$cohortDefinitionSet$cohortName,
-                              " with ids: ", r_toAdd$cohortDefinitionSet$cohortId,
+      ParallelLogger::logInfo("[Operate cohorts] Creating cohorts: ", r_cohortDefinitionSetToAdd$cohortDefinitionSet$cohortName,
+                              " with ids: ", r_cohortDefinitionSetToAdd$cohortDefinitionSet$cohortId,
                               " to database ", input$selectDatabases_pickerInput)
 
     })
@@ -259,12 +233,11 @@ mod_operateCohorts_server <- function(id, r_connectionHandlers, r_workbench) {
     #
     # evaluate the cohorts to append; if accepted increase output to trigger closing actions
     #
-    r_append_accepted_counter <- mod_appendCohort_server("matchCohort", r_connectionHandlers, r_workbench, r_toAdd )
+    rf_append_accepted_counter <- mod_fct_appendCohort_server("matchCohort", r_connectionHandler, r_cohortDefinitionSetToAdd )
 
     # close and reset
-    shiny::observeEvent(r_append_accepted_counter(), {
+    shiny::observeEvent(rf_append_accepted_counter(), {
       # change in r_workbench$cohortsSummaryDatabases will update output$selectDatabases_pickerInput_uiOutput <- shiny::renderUI({
-      # this will chain update the rest
       r$cohortDefinitionSet <- NULL
     })
   })
@@ -280,7 +253,7 @@ mod_operateCohorts_server <- function(id, r_connectionHandlers, r_workbench) {
   cohortNames <- cohortOverlap |> dplyr::select(-numberOfSubjects) |> colnames()  |> setdiff("newset")
 
   cohortOverlapPlot  <- cohortOverlap |>
-     # convert true false to cohort names
+    # convert true false to cohort names
     dplyr::mutate(dplyr::across(dplyr::all_of(cohortNames), ~ifelse(., as.character(dplyr::cur_column()), '')))  |>
     # join cohort names
     dplyr::mutate(cohort_vector = stringr::str_c( !!!rlang::syms(cohortNames), sep = " ")) |>