+++ tests/testmanual/manualtest-mod_SelectDatabases.R
@@ -1,0 +1,31 @@
+# build parameters --------------------------------------------------------------
+devtools::load_all(".")
+source(testthat::test_path("setup.R"))
+source(testthat::test_path("helper.R"))
+
+fcr_setUpLogger()
+
+r_connectionHandler <- shiny::reactiveValues(
+  cohortTableHandler = NULL,
+  hasChangeCounter = 0
+)
+
+
+# run module --------------------------------------------------------------
+devtools::load_all(".")
+
+app <- shiny::shinyApp(
+  shiny::fluidPage(
+    mod_selectDatabases_ui("select_configuration")
+  ),
+  function(input,output,session){
+    mod_selectDatabases_server("select_configuration", databasesConfig, r_connectionHandler)
+  },
+  options = list(launch.browser=TRUE)
+)
+
+
+app
+
+
+