+++ NEWS.md
@@ -1,3 +1,11 @@
+# CohortOperations2 1.0.0
+- Major release
+- Analyses moved to CO2AnalysisModules
+- removed renv
+- downgraded features to make it simpler at the moment
+  - connect to only one database
+  - run analysis on main process instead of icp
+
 # CohortOperations2 0.3.0
 
 - Added Cohorts Overlaps
@@ -15,4 +23,3 @@
 # CohortOperations2 0.2.1
 
 - Fixed operations with Atlas cohorts
-f